Add interface.py and use it
This adds interface.py, which is a file that serves to wrap all calls
made to or from the discord.py library so that it can be mocked for the
bot's internal logic.

Additionally it also makes things more modular since replying to a
message can be done with just a call to interface.reply, which prevents
the need for copy-pasting.