Add history cog
This cog is for storing all messages the bot can see in a local SQLite
database. "History" is an apt way of describing this database that's
meant to contain basically everything.

For future-proofing, we store the raw JSON of all messages (the raw
JSON being what the Discord gateway sends for every message). That way,
if there's new data (such as forwarded messages), it will be stored
without needing new code written to explicity handle it.