Interdict the create_message function
This function is an internal discord.py function that gets called
whenever a message is fetched, or when the bot sends a message.

This is mainly needed to get the raw JSON of a fetched message, since
discord.py does not store the raw JSON and discards it once the
discord.Message object is created.