history: Implement message updates
This adds a new column, 'version', to store which version a message is.
Whenever a message is updated (e.g. the content is edited), the new
version of the message is added to the table with an incremented
version. The old version is kept in the table.

This also implements it in the create_message hook.